<fix> correcao validateToken home

+++ lista-de-compras/src/app/pages/home/home.component.ts
@@ -8,32 +8,33 @@ import { ToastrService } from 'ngx-toastr';
   standalone: true,
   imports: [],
   templateUrl: './home.component.html',
-  styleUrl: './home.component.scss'
+  styleUrl: './home.component.scss',
 })
 export class HomeComponent implements OnInit {
-
   constructor(
     private TokenService: TokenService,
     private router: Router,
     private toast: ToastrService
-  ){
-  }
+  ) {}
 
   ngOnInit(): void {
-    console.log("Iniciou a homepage.")
-    this.TokenService.validateToken().subscribe(
-      res => {
-        if(!res){
+    console.log('Iniciou a homepage.');
+    if (!localStorage.getItem('token')) {
+      console.log('Token não encontrado. Redirecionando para o login.');
+      this.router.navigate(['login']);
+    } else {
+      this.TokenService.validateToken().subscribe((res) => {
+        if (!res) {
           localStorage.clear();
-          console.log("Sessão de usuário expiradar. Fazer fazer login novamente.");
-          this.toast.info("Sua sessão expirou");
-          this.router.navigate(["login"]);
+          console.log(
+            'Sessão de usuário expiradar. Fazer fazer login novamente.'
+          );
+          this.toast.info('Sua sessão expirou');
+          this.router.navigate(['login']);
         } else {
-          console.log("Token está válido.")
+          console.log('Token está válido.');
         }
-      }
-    );
+      });
+    }
   }
-
-
 }
